move State back to a renderer-state module
