+++ module/parser-base/build.gradle.kts
@@ -1,8 +1,5 @@
-import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
-
 plugins {
     id("godtools-shared.module-conventions")
-    kotlin("plugin.parcelize")
 }
 
 android {
@@ -12,25 +9,12 @@ android {
 kotlin {
     configureJsTargets()
 
-    androidTarget {
-        @OptIn(ExperimentalKotlinGradlePluginApi::class)
-        compilerOptions.freeCompilerArgs.addAll(
-            "-P",
-            "plugin:org.jetbrains.kotlin.parcelize:additionalAnnotation=" +
-                "org.cru.godtools.shared.tool.state.internal.Parcelize",
-        )
-    }
-
     sourceSets {
         val commonMain by getting {
             dependencies {
+                api(project(":module:renderer-state"))
+
                 implementation(libs.androidx.annotation)
-                implementation(libs.kotlin.coroutines.core)
-            }
-        }
-        val commonTest by getting {
-            dependencies {
-                implementation(libs.turbine)
             }
         }
     }
