--- module/parser-base/src/androidMain/kotlin/org/cru/godtools/shared/tool/state/internal/Parcelable.kt
@@ -1,3 +1,0 @@
-package org.cru.godtools.shared.tool.state.internal
-
-actual typealias Parcelable = android.os.Parcelable
