--- module/parser-base/src/commonMain/kotlin/org/cru/godtools/shared/tool/state/internal/Parcelable.kt
@@ -1,7 +1,0 @@
-package org.cru.godtools.shared.tool.state.internal
-
-expect interface Parcelable
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.BINARY)
-internal annotation class Parcelize
