+++ module/parser-base/src/commonTest/kotlin/org/cru/godtools/shared/tool/parser/model/EventIdTest.kt
@@ -5,7 +5,7 @@ import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue
-import org.cru.godtools.shared.tool.state.State
+import org.cru.godtools.shared.renderer.state.State
 
 class EventIdTest {
     companion object {
