--- module/parser-base/src/iosMain/kotlin/org/cru/godtools/shared/tool/state/internal/Parcelable.kt
@@ -1,3 +1,0 @@
-package org.cru.godtools.shared.tool.state.internal
-
-actual interface Parcelable
