+++ module/parser-expressions/build.gradle.kts
@@ -13,7 +13,7 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                api(project(":module:parser-base"))
+                api(project(":module:renderer-state"))
 
                 implementation(kotlin("stdlib-common"))
                 implementation(libs.antlr.kotlin.runtime)
