+++ module/parser-expressions/src/commonMain/kotlin/org/cru/godtools/shared/tool/parser/expressions/Expression.kt
@@ -7,10 +7,10 @@ import org.antlr.v4.kotlinruntime.ParserRuleContext
 import org.antlr.v4.kotlinruntime.Token
 import org.antlr.v4.kotlinruntime.misc.ParseCancellationException
 import org.antlr.v4.kotlinruntime.tree.TerminalNode
+import org.cru.godtools.shared.renderer.state.State
 import org.cru.godtools.shared.tool.parser.expressions.grammar.StateExpressionEvaluator
 import org.cru.godtools.shared.tool.parser.expressions.grammar.generated.StateExpressionLexer
 import org.cru.godtools.shared.tool.parser.expressions.grammar.generated.StateExpressionParser
-import org.cru.godtools.shared.tool.state.State
 
 class Expression internal constructor(
     private val expr: StateExpressionParser.BooleanExprContext?,
