+++ module/parser/src/commonMain/kotlin/org/cru/godtools/shared/tool/parser/model/Visibility.kt
@@ -7,11 +7,11 @@ import kotlin.js.ExperimentalJsExport
 import kotlin.js.JsExport
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
+import org.cru.godtools.shared.renderer.state.State
 import org.cru.godtools.shared.tool.parser.expressions.Expression
 import org.cru.godtools.shared.tool.parser.expressions.toExpressionOrNull
 import org.cru.godtools.shared.tool.parser.util.FlowWatcher.Companion.watch
 import org.cru.godtools.shared.tool.parser.xml.XmlPullParser
-import org.cru.godtools.shared.tool.state.State
 
 private const val XML_INVISIBLE_IF = "invisible-if"
 private const val XML_GONE_IF = "gone-if"
