+++ module/renderer-state/src/androidMain/kotlin/org/cru/godtools/shared/renderer/state/internal/Parcelable.kt
@@ -1,0 +1,3 @@
+package org.cru.godtools.shared.renderer.state.internal
+
+actual typealias Parcelable = android.os.Parcelable
