+++ module/renderer-state/src/commonMain/kotlin/org/cru/godtools/shared/renderer/state/internal/Parcelable.kt
@@ -1,0 +1,7 @@
+package org.cru.godtools.shared.renderer.state.internal
+
+expect interface Parcelable
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.BINARY)
+internal annotation class Parcelize
