+++ module/renderer-state/src/jsMain/kotlin/org/cru/godtools/shared/renderer/state/internal/Parcelable.kt
@@ -1,0 +1,3 @@
+package org.cru.godtools.shared.renderer.state.internal
+
+actual interface Parcelable
